Separate node definition from entity definition
